fix(search): address code review findings in scroll-to-term feature
- Extract SEARCH_QUERY_STATE_KEY and getNavigationSearchQuery to
  lib/searchNavigationState.ts so writer (SearchResultCard) and reader
  (useScrollToSearchTerm) share the same constant
- Extract waitUntilHeightStabilizes from scrollToHeadlineHash to module
  scope with proper cancellation support; use it in useScrollToSearchTerm
  instead of a flat 500ms timeout to avoid racing slow-rendering content
- Replace dead scrollContainer guard with actual usage: pass it to
  waitUntilHeightStabilizes for height-stability polling
- Track highlighted visit keys in a ref (visitKey + searchQuery) to
  prevent the highlight re-firing when the user navigates Back
