+++ ui/leafwiki-ui/src/features/search/SearchResultCard.tsx
@@ -1,6 +1,7 @@
 import { SearchResultItem } from '@/lib/api/search'
 import { createNavigationVisitState } from '@/lib/navigationVisit'
 import { buildViewUrl } from '@/lib/routePath'
+import { SEARCH_QUERY_STATE_KEY } from '@/lib/searchNavigationState'
 import { normalizeWikiRoutePath } from '@/lib/wikiPath'
 import { forwardRef } from 'react'
 import { Link, useLocation } from 'react-router-dom'
@@ -37,7 +38,7 @@ const SearchResultCard = forwardRef<HTMLAnchorElement, SearchResultCardProps>(
       <Link
         ref={ref}
         to={resultUrl}
-        state={createNavigationVisitState({ leafwikiSearchQuery: searchQuery })}
+        state={createNavigationVisitState({ [SEARCH_QUERY_STATE_KEY]: searchQuery })}
         data-testid={`search-result-card-${item.page_id}`}
         aria-current={isRouteActive ? 'page' : undefined}
         onMouseEnter={onMouseEnter}
