+++ ui/leafwiki-ui/src/lib/scrollToHeadline.ts
@@ -3,6 +3,48 @@ type ScrollToHeadlineOptions = {
   waitForStableLayout?: boolean
 }
 
+export function waitUntilHeightStabilizes(
+  element: HTMLElement,
+  callback: () => void,
+  interval = 250,
+  maxTotalTime = 3000,
+  stableTime = 500,
+): () => void {
+  let lastHeight = element.scrollHeight
+  let stableFor = 0
+  let elapsedTime = 0
+  let cancelled = false
+  let currentTimer: ReturnType<typeof setTimeout>
+
+  const checkHeight = () => {
+    if (cancelled) return
+    const currentHeight = element.scrollHeight
+    if (currentHeight === lastHeight) {
+      stableFor += interval
+      if (stableFor >= stableTime) {
+        callback()
+        return
+      }
+    } else {
+      lastHeight = currentHeight
+      stableFor = 0
+    }
+    elapsedTime += interval
+    if (elapsedTime < maxTotalTime) {
+      currentTimer = setTimeout(checkHeight, interval)
+    } else {
+      callback()
+    }
+  }
+
+  currentTimer = setTimeout(checkHeight, interval)
+
+  return () => {
+    cancelled = true
+    clearTimeout(currentTimer)
+  }
+}
+
 export function scrollToHeadlineHash(
   hash: string,
   {
@@ -14,40 +56,6 @@ export function scrollToHeadlineHash(
     'scroll-container',
   ) as HTMLElement | null
   if (!contentContainer) return
-
-  function waitUntilHeightStabilizes(
-    element: HTMLElement,
-    callback: () => void,
-    interval = 250,
-    maxTotalTime = 3000,
-    stableTime = 500,
-  ) {
-    let lastHeight = element.scrollHeight
-    let stableFor = 0
-    let elapsedTime = 0
-
-    const checkHeight = () => {
-      const currentHeight = element.scrollHeight
-      if (currentHeight === lastHeight) {
-        stableFor += interval
-        if (stableFor >= stableTime) {
-          callback()
-          return
-        }
-      } else {
-        lastHeight = currentHeight
-        stableFor = 0
-      }
-      elapsedTime += interval
-      if (elapsedTime < maxTotalTime) {
-        setTimeout(checkHeight, interval)
-      } else {
-        callback()
-      }
-    }
-
-    setTimeout(checkHeight, interval)
-  }
 
   const scrollToTarget = () => {
     const rawHeadlineId = hash.substring(1)
